prueba2019348293423

+++ Terato/segdataset.py
@@ -141,6 +141,5 @@ class SegmentationDataset(VisionDataset):
 
             except:
                 sample['masks'].append(self.transforms(black_mask))
-        sample['image'] = sample['image'].long()
-        sample['masks'] = torch.cat(sample['masks'], 0).long() # Concatenate the masks in one unique tensor
+        sample['masks'] = torch.cat(sample['masks'], 0).float() # Concatenate the masks in one unique tensor
         return sample
